# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Solunex Lab Backend",
                 "version":  "0.1.0"
             },
    "paths":  {
...
                  "/api/templates":  {
                                         "post":  {
                                                      "tags":  [
                                                                   "templates"
                                                               ],
                                                      "summary":  "Create Template",
                                                      "operationId":  "create_template_api_templates_post",
                                                      "requestBody":  {
                                                                          "content":  {
                                                                                          "application/json":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/TestTemplateCreate"
                                                                                                                              }
                                                                                                               }
                                                                                      },
                                                                          "required":  true
                                                                      },
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "Successful Response",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/TestTemplateOut"
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                },
                                                                        "422":  {
                                                                                    "description":  "Validation Error",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                }
                                                                    }
                                                  }
                                     },
                  "/api/templates/active":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "templates"
                                                                     ],
                                                            "summary":  "List Active Templates",
                                                            "operationId":  "list_active_templates_api_templates_active_get",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "test_type_id",
                                                                                   "in":  "query",
                                                                                   "required":  false,
                                                                                   "schema":  {
                                                                                                  "anyOf":  [
                                                                                                                {
                                                                                                                    "type":  "integer"
                                                                                                                },
                                                                                                                {
                                                                                                                    "type":  "null"
                                                                                                                }
                                                                                                            ],
                                                                                                  "title":  "Test Type Id"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Successful Response",
                                                                                          "content":  {
                                                                                                          "application/json":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "type":  "array",
                                                                                                                                                  "items":  {
                                                                                                                                                                "$ref":  "#/components/schemas/TestTemplateOut"
                                                                                                                                                            },
                                                                                                                                                  "title":  "Response List Active Templates Api Templates Active Get"
                                                                                                                                              }
                                                                                                                               }
                                                                                                      }
                                                                                      },
                                                                              "422":  {
                                                                                          "description":  "Validation Error",
                                                                                          "content":  {
                                                                                                          "application/json":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                              }
                                                                                                                               }
                                                                                                      }
                                                                                      }
                                                                          }
                                                        }
                                            },
                  "/api/templates/{template_id}":  {
                                                       "get":  {
                                                                   "tags":  [
                                                                                "templates"
                                                                            ],
                                                                   "summary":  "Get Template",
                                                                   "operationId":  "get_template_api_templates__template_id__get",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "template_id",
                                                                                          "in":  "path",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "integer",
                                                                                                         "title":  "Template Id"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Successful Response",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/TestTemplateOut"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             },
                                                                                     "422":  {
                                                                                                 "description":  "Validation Error",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             }
                                                                                 }
                                                               }
                                                   },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "TestTemplateCreate":  {
                                                                  "properties":  {
                                                                                     "test_type_id":  {
                                                                                                          "type":  "integer",
                                                                                                          "title":  "Test Type Id"
                                                                                                      },
                                                                                     "title":  {
                                                                                                   "type":  "string",
                                                                                                   "maxLength":  255,
                                                                                                   "minLength":  1,
                                                                                                   "title":  "Title"
                                                                                               },
                                                                                     "structure":  {
                                                                                                       "additionalProperties":  true,
                                                                                                       "type":  "object",
                                                                                                       "title":  "Structure"
                                                                                                   },
                                                                                     "rules":  {
                                                                                                   "anyOf":  [
                                                                                                                 {
                                                                                                                     "additionalProperties":  true,
                                                                                                                     "type":  "object"
                                                                                                                 },
                                                                                                                 {
                                                                                                                     "type":  "null"
                                                                                                                 }
                                                                                                             ],
                                                                                                   "title":  "Rules"
                                                                                               },
                                                                                     "is_active":  {
                                                                                                       "type":  "boolean",
                                                                                                       "title":  "Is Active",
                                                                                                       "default":  true
                                                                                                   }
                                                                                 },
                                                                  "type":  "object",
                                                                  "required":  [
                                                                                   "test_type_id",
                                                                                   "title",
                                                                                   "structure"
                                                                               ],
                                                                  "title":  "TestTemplateCreate"
                                                              },
                                       "TestTemplateOut":  {
                                                               "properties":  {
                                                                                  "id":  {
                                                                                             "type":  "integer",
                                                                                             "title":  "Id"
                                                                                         },
                                                                                  "test_type_id":  {
                                                                                                       "type":  "integer",
                                                                                                       "title":  "Test Type Id"
                                                                                                   },
                                                                                  "title":  {
                                                                                                "type":  "string",
                                                                                                "title":  "Title"
                                                                                            },
                                                                                  "structure":  {
                                                                                                    "additionalProperties":  true,
                                                                                                    "type":  "object",
                                                                                                    "title":  "Structure"
                                                                                                },
                                                                                  "rules":  {
                                                                                                "anyOf":  [
                                                                                                              {
                                                                                                                  "additionalProperties":  true,
                                                                                                                  "type":  "object"
                                                                                                              },
                                                                                                              {
                                                                                                                  "type":  "null"
                                                                                                              }
                                                                                                          ],
                                                                                                "title":  "Rules"
                                                                                            },
                                                                                  "is_active":  {
                                                                                                    "type":  "boolean",
                                                                                                    "title":  "Is Active"
                                                                                                },
                                                                                  "created_at":  {
                                                                                                     "type":  "string",
                                                                                                     "format":  "date-time",
                                                                                                     "title":  "Created At"
                                                                                                 },
                                                                                  "updated_at":  {
                                                                                                     "type":  "string",
                                                                                                     "format":  "date-time",
                                                                                                     "title":  "Updated At"
                                                                                                 }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "id",
                                                                                "test_type_id",
                                                                                "title",
                                                                                "structure",
                                                                                "rules",
                                                                                "is_active",
                                                                                "created_at",
                                                                                "updated_at"
                                                                            ],
                                                               "title":  "TestTemplateOut"
                                                           },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           }
                                   }
                   }
}